requirements.txt
tqdm==4.33.0
urllib3==1.24.3
requests==2.25.1
python-dotenv==0.19.2
